+++ CHANGELOG.md
@@ -11,6 +11,7 @@
     * MediaTrait: Use PSR-16 cache
     * Improved `Utils::normalizePath()` to support non-protocol URLs
 1. [](#bugfix)
+    * Fixed issue with `Truncator::truncateWords` and `Truncator::truncateLetters` when string not wrapped in tags [#2432](https://github.com/getgrav/grav/issues/2432)
     * Fixed `Undefined method closure::fields()` when getting avatar for user, thanks @Romarain [#2422](https://github.com/getgrav/grav/issues/2422)
     * Grav 1.6: Added `Flex[Class]::getFlexType()` to all flex classes to have the interface for all Flex classes
     * FlexObjects: Remove null values during save